refactor: js 파일 모듈로 분리
