+++ src/main/java/chess/WebUIChessApplication.java
@@ -23,7 +23,7 @@ public class WebUIChessApplication {
 
     private static final Gson gson = new Gson();
 
-    public static void main(String[] args) {
+    public static  void main(String[] args) {
         Board board = new Board();
         ChessGame chessGame = new ChessGame(board, new Turn(Team.WHITE));
