+++ src/main/resources/public/assets/css/index.css
@@ -148,7 +148,7 @@ section {
 .result p {
     position: relative;
     top: 40%;
-    font-size: xxx-large;
+    font-size: xx-large;
     font-family: "pencil", sans-serif;
     color: #eee;
 }
