+++ src/main/resources/templates/index.html
@@ -23,6 +23,6 @@
     <img class="blackTurn" src="assets/images/black_turn.png">
   </div>
 </section>
-<script type="text/javascript" src="/assets/js/index.js"></script>
+<script type="module" src="/assets/js/index.js"></script>
 </body>
 </html>
